Fix: Next mech was skipped on saving mech
The bug was: When a mech died or was saved, the next mech would flicker
for one frame (i.e., be skipped during rendering) and would also be
ignored during the physics update.

Reason: When you iterate as follows:

    for x in myList:
        if foo(x):
            myList.remove(x)

...you will skip elements because the iteration does not see the
removal. Rewritten as list comprehension. The downside is that the
predicate of the list comprehension mutates the elements (calls act() on
the mechs).